forex/requirements.txt
faker
kafka-python

